mark an item as done

+++ src/components/Item.tsx
@@ -1,27 +1,47 @@
 import React from 'react';
+import type { ItemType } from '../App';
+
+interface OptionalHandlers {
+  item: ItemType;
+  onDelete?: (id: string) => void;
+  onDone?: (item: ItemType) => void;
+}
+
+Item.defaultProps = {
+  onDelete: undefined,
+  onDone: undefined,
+};
 
 function Item({
   item,
   onDelete,
-}: {
-  item: {
-    id: string;
-    task: string;
-    status: string;
-    date: Date;
-  };
-  onDelete(id: string): void;
-}): React.ReactElement {
+  onDone,
+}: OptionalHandlers): React.ReactElement {
+  if (onDelete && onDone) {
+    return (
+      <div>
+        {item.id} {item.task} {item.status} {item.date.getTime()}
+        <button
+          className="ml-4 bg-gray-400"
+          type="button"
+          onClick={(): void => onDelete(item.id)}
+        >
+          Delete
+        </button>
+        <button
+          className="ml-4 bg-gray-400"
+          type="button"
+          onClick={(): void => onDone(item)}
+        >
+          Done
+        </button>
+      </div>
+    );
+  }
+
   return (
     <div>
       {item.id} {item.task} {item.status} {item.date.getTime()}
-      <button
-        className="ml-4 bg-gray-400"
-        type="button"
-        onClick={(): void => onDelete(item.id)}
-      >
-        Delete
-      </button>
     </div>
   );
 }
